Moved TunnelExitSocket/DataChecker to separate file
