+++ ipv8/test/messaging/anonymization/test_exit_socket.py
@@ -2,7 +2,8 @@ from binascii import unhexlify
 from unittest.mock import Mock
 
 from ....keyvault.private.libnaclkey import LibNaCLSK
-from ....messaging.anonymization.tunnel import PEER_FLAG_EXIT_BT, PEER_FLAG_EXIT_IPV8, TunnelExitSocket
+from ....messaging.anonymization.tunnel import PEER_FLAG_EXIT_BT, PEER_FLAG_EXIT_IPV8
+from ....messaging.anonymization.exit_socket import TunnelExitSocket
 from ....peer import Peer
 from ...base import TestBase
 from .test_datachecker import dht_pkt, ipv8_pkt, tracker_pkt, tunnel_pkt, utp_pkt
@@ -17,7 +18,8 @@ class TestExitSocket(TestBase):
         """
         Check if the ExitSocket correctly detects forbidden packets.
         """
-        overlay = Mock(_prefix=unhexlify('000281ded07332bdc775aa5a46f96de9f8f390bbc9f3'))
+        get_prefix = Mock(return_value=unhexlify('000281ded07332bdc775aa5a46f96de9f8f390bbc9f3'))
+        overlay = Mock(get_prefix=get_prefix)
         exit_socket = TunnelExitSocket(0, Mock(peer=Peer(LibNaCLSK(b"\x00" * 64))), overlay)
 
         overlay.settings.peer_flags = {}
